Separating out the interactive and regular features of Telegram
