+++ lux/test/unit/lux/lenses/telegram/telegram_bot_lens_test.exs
@@ -1,23 +1,26 @@
 defmodule Lux.Lenses.TelegramBotLensTest do
   use UnitAPICase, async: true
 
+  # Note: Tests for interactive features (polls, games, live location, stickers, inline queries)
+  # have been moved to Lux.Lenses.TelegramInteractiveFeaturesTest
+
   alias Lux.Lenses.TelegramBotLens
 
   setup do
     Req.Test.verify_on_exit!()
 
     # Set up test token
-    original_config = Application.get_env(:lux, :api_keys)
+      original_config = Application.get_env(:lux, :api_keys)
 
-    on_exit(fn ->
-      Application.put_env(:lux, :api_keys, original_config)
-    end)
+      on_exit(fn ->
+        Application.put_env(:lux, :api_keys, original_config)
+      end)
 
-    # Set both telegram_bot and integration_telegram_bot keys
-    Application.put_env(:lux, :api_keys, [
-      telegram_bot: "test_token",
-      integration_telegram_bot: "test_integration_token"
-    ])
+      # Set both telegram_bot and integration_telegram_bot keys
+      Application.put_env(:lux, :api_keys, [
+        telegram_bot: "test_token",
+        integration_telegram_bot: "test_integration_token"
+      ])
 
     :ok
   end
@@ -86,11 +89,11 @@ defmodule Lux.Lenses.TelegramBotLensTest do
           "text" => "Hello, world!"
         }
       }} = TelegramBotLens.focus(%{
-        method: "sendMessage",
-        chat_id: "123456789",
+          method: "sendMessage",
+          chat_id: "123456789",
         text: "Hello, world!",
         token: "test_integration_token"
-      })
+        })
     end
 
     test "focus with sendPhoto method" do
@@ -127,12 +130,12 @@ defmodule Lux.Lenses.TelegramBotLensTest do
           "photo" => [%{"file_id" => "photo1"} | _]
         }
       }} = TelegramBotLens.focus(%{
-        method: "sendPhoto",
-        chat_id: "123456789",
-        photo: "https://example.com/image.jpg",
+          method: "sendPhoto",
+          chat_id: "123456789",
+          photo: "https://example.com/image.jpg",
         caption: "Test photo",
         token: "test_integration_token"
-      })
+        })
     end
 
     test "focus with sendDocument method" do
@@ -171,12 +174,12 @@ defmodule Lux.Lenses.TelegramBotLensTest do
           "document" => %{"file_id" => "doc1"}
         }
       }} = TelegramBotLens.focus(%{
-        method: "sendDocument",
-        chat_id: "123456789",
-        document: "https://example.com/file.pdf",
+          method: "sendDocument",
+          chat_id: "123456789",
+          document: "https://example.com/file.pdf",
         caption: "Test document",
         token: "test_integration_token"
-      })
+        })
     end
 
     test "handles error response" do
@@ -191,11 +194,11 @@ defmodule Lux.Lenses.TelegramBotLensTest do
         "ok" => false,
         "description" => "Bad Request: chat not found"
       }} = TelegramBotLens.focus(%{
-        method: "sendMessage",
-        chat_id: "invalid",
+          method: "sendMessage",
+          chat_id: "invalid",
         text: "Hello, world!",
         token: "test_integration_token"
-      })
+        })
     end
 
     test "handles unexpected response format" do
@@ -210,375 +213,6 @@ defmodule Lux.Lenses.TelegramBotLensTest do
         })
     end
 
-    test "send_quiz/5" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "sendPoll")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["chat_id"] == "123456789"
-        assert body["question"] == "What is Elixir?"
-        assert body["options"] == ["A language", "A framework", "A database"]
-        assert body["type"] == "quiz"
-        assert body["correct_option_id"] == 0
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "message_id" => 1,
-            "from" => %{"id" => 987654321, "is_bot" => true},
-            "chat" => %{"id" => 123456789},
-            "poll" => %{
-              "id" => "quiz123",
-              "question" => "What is Elixir?",
-              "options" => [
-                %{"text" => "A language", "voter_count" => 0},
-                %{"text" => "A framework", "voter_count" => 0},
-                %{"text" => "A database", "voter_count" => 0}
-              ],
-              "type" => "quiz",
-              "correct_option_id" => 0
-            }
-          }
-        })
-      end)
-
-      assert {:ok, %{"poll" => %{"id" => "quiz123"}}} =
-        TelegramBotLens.send_quiz(
-          "123456789",
-          "What is Elixir?",
-          ["A language", "A framework", "A database"],
-          0
-        )
-    end
-
-    test "stop_poll/3" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "stopPoll")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["chat_id"] == "123456789"
-        assert body["message_id"] == 42
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "id" => "poll123",
-            "question" => "Question",
-            "is_closed" => true
-          }
-        })
-      end)
-
-      assert {:ok, %{"id" => "poll123", "is_closed" => true}} =
-        TelegramBotLens.stop_poll("123456789", 42)
-    end
-
-    test "send_game/3" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "sendGame")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["chat_id"] == "123456789"
-        assert body["game_short_name"] == "test_game"
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "message_id" => 1,
-            "from" => %{"id" => 987654321, "is_bot" => true},
-            "chat" => %{"id" => 123456789},
-            "game" => %{
-              "title" => "Test Game",
-              "description" => "A test game",
-              "photo" => [%{"file_id" => "game_photo"}]
-            }
-          }
-        })
-      end)
-
-      assert {:ok, %{"game" => %{"title" => "Test Game"}}} =
-        TelegramBotLens.send_game("123456789", "test_game")
-    end
-
-    test "set_game_score/3" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "setGameScore")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["user_id"] == 123456789
-        assert body["score"] == 100
-        assert body["chat_id"] == "987654321"
-        assert body["message_id"] == 42
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "message_id" => 42,
-            "from" => %{"id" => 987654321, "is_bot" => true},
-            "game" => %{"title" => "Test Game", "score" => 100}
-          }
-        })
-      end)
-
-      assert {:ok, %{"game" => %{"score" => 100}}} =
-        TelegramBotLens.set_game_score(123456789, 100, %{
-          chat_id: "987654321",
-          message_id: 42
-        })
-    end
-
-    test "get_game_high_scores/2" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "getGameHighScores")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["user_id"] == 123456789
-        assert body["chat_id"] == "987654321"
-        assert body["message_id"] == 42
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => [
-            %{
-              "position" => 1,
-              "user" => %{"id" => 123456789, "first_name" => "Player 1"},
-              "score" => 100
-            }
-          ]
-        })
-      end)
-
-      assert {:ok, [%{"score" => 100}]} =
-        TelegramBotLens.get_game_high_scores(123456789, %{
-          chat_id: "987654321",
-          message_id: 42
-        })
-    end
-
-    test "send_live_location/5" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "sendLocation")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["chat_id"] == "123456789"
-        assert body["latitude"] == 37.7749
-        assert body["longitude"] == -122.4194
-        assert body["live_period"] == 60
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "message_id" => 1,
-            "from" => %{"id" => 987654321, "is_bot" => true},
-            "chat" => %{"id" => 123456789},
-            "location" => %{
-              "latitude" => 37.7749,
-              "longitude" => -122.4194,
-              "live_period" => 60
-            }
-          }
-        })
-      end)
-
-      assert {:ok, %{"location" => %{"live_period" => 60}}} =
-        TelegramBotLens.send_live_location("123456789", 37.7749, -122.4194, 60)
-    end
-
-    test "edit_live_location/5" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "editMessageLiveLocation")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["chat_id"] == "123456789"
-        assert body["message_id"] == 42
-        assert body["latitude"] == 37.7750
-        assert body["longitude"] == -122.4195
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "message_id" => 42,
-            "location" => %{
-              "latitude" => 37.7750,
-              "longitude" => -122.4195
-            }
-          }
-        })
-      end)
-
-      assert {:ok, %{"location" => %{"latitude" => 37.7750}}} =
-        TelegramBotLens.edit_live_location("123456789", 42, 37.7750, -122.4195)
-    end
-
-    test "stop_live_location/3" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "stopMessageLiveLocation")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["chat_id"] == "123456789"
-        assert body["message_id"] == 42
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "message_id" => 42,
-            "location" => %{
-              "latitude" => 37.7749,
-              "longitude" => -122.4194,
-              "live_period" => 0
-            }
-          }
-        })
-      end)
-
-      assert {:ok, %{"location" => %{"live_period" => 0}}} =
-        TelegramBotLens.stop_live_location("123456789", 42)
-    end
-
-    test "get_sticker_set/1" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "getStickerSet")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["name"] == "test_set"
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => %{
-            "name" => "test_set",
-            "title" => "Test Sticker Set",
-            "sticker_type" => "regular",
-            "stickers" => [
-              %{
-                "file_id" => "sticker1",
-                "width" => 512,
-                "height" => 512,
-                "is_animated" => false
-              }
-            ]
-          }
-        })
-      end)
-
-      assert {:ok, %{"name" => "test_set", "stickers" => [%{"file_id" => "sticker1"}]}} =
-        TelegramBotLens.get_sticker_set("test_set")
-    end
-
-    test "create_article_result/4" do
-      input_content = TelegramBotLens.create_text_content("Test message")
-      result = TelegramBotLens.create_article_result("1", "Test Title", input_content, %{
-        description: "Test description"
-      })
-
-      assert result.type == "article"
-      assert result.id == "1"
-      assert result.title == "Test Title"
-      assert result.description == "Test description"
-      assert result.input_message_content == input_content
-    end
-
-    test "create_photo_result/4" do
-      result = TelegramBotLens.create_photo_result(
-        "1",
-        "https://example.com/photo.jpg",
-        "https://example.com/thumb.jpg",
-        %{title: "Test Photo"}
-      )
-
-      assert result.type == "photo"
-      assert result.id == "1"
-      assert result.photo_url == "https://example.com/photo.jpg"
-      assert result.thumbnail_url == "https://example.com/thumb.jpg"
-      assert result.title == "Test Photo"
-    end
-
-    test "answer_inline_query/3" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "answerInlineQuery")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["inline_query_id"] == "query123"
-        assert is_list(body["results"])
-        assert length(body["results"]) == 1
-        assert hd(body["results"])["type"] == "article"
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => true
-        })
-      end)
-
-      text_content = TelegramBotLens.create_text_content("Test message")
-      result = TelegramBotLens.create_article_result("1", "Test Title", text_content)
-
-      assert {:ok, true} =
-        TelegramBotLens.answer_inline_query("query123", [result], %{cache_time: 300})
-    end
-
-    test "set_inline_handler/2" do
-      handler = fn query ->
-        text_content = TelegramBotLens.create_text_content("You searched for: #{query["query"]}")
-        [TelegramBotLens.create_article_result("1", "Result", text_content)]
-      end
-
-      TelegramBotLens.set_inline_handler(handler, %{cache_time: 300})
-
-      assert {^handler, %{cache_time: 300}} = Process.get(:telegram_inline_handler)
-    end
-
-    test "process_inline_query/1 with handler set" do
-      Req.Test.expect(Lux.Lens, fn conn ->
-        assert conn.method == "POST"
-        assert String.ends_with?(conn.request_path, "answerInlineQuery")
-
-        body = Jason.decode!(conn.adapter |> elem(1) |> Map.get(:req_body))
-        assert body["inline_query_id"] == "query123"
-        assert is_list(body["results"])
-        assert length(body["results"]) == 1
-        [result] = body["results"]
-        assert result["type"] == "article"
-        assert result["input_message_content"]["message_text"] == "You searched for: test"
-
-        Req.Test.json(conn, %{
-          "ok" => true,
-          "result" => true
-        })
-      end)
-
-      handler = fn query ->
-        text_content = TelegramBotLens.create_text_content("You searched for: #{query["query"]}")
-        [TelegramBotLens.create_article_result("1", "Result", text_content)]
-      end
-
-      TelegramBotLens.set_inline_handler(handler, %{cache_time: 300})
-
-      assert {:ok, true} =
-        TelegramBotLens.process_inline_query(%{
-          "id" => "query123",
-          "from" => %{"id" => 123456789},
-          "query" => "test"
-        })
-    end
-
-    test "process_inline_query/1 without handler" do
-      Process.delete(:telegram_inline_handler)
-
-      assert {:error, "No inline query handler set"} =
-        TelegramBotLens.process_inline_query(%{
-          "id" => "query123",
-          "query" => "test"
-        })
-    end
 
     test "set_webhook/2" do
       Req.Test.expect(Lux.Lens, fn conn ->
@@ -1066,10 +700,10 @@ defmodule Lux.Lenses.TelegramBotLensTest do
           %{"file_id" => "photo2", "file_size" => 5678, "height" => 600, "width" => 800}
         ]
       }} = TelegramBotLens.send_photo(
-        "123456789",
-        "https://example.com/image.jpg",
-        %{caption: "Beautiful sunset", parse_mode: "HTML"}
-      )
+          "123456789",
+          "https://example.com/image.jpg",
+          %{caption: "Beautiful sunset", parse_mode: "HTML"}
+        )
     end
 
     test "send_document/3" do
@@ -1118,10 +752,10 @@ defmodule Lux.Lenses.TelegramBotLensTest do
         },
         "message_id" => 1
       }} = TelegramBotLens.send_document(
-        "123456789",
-        "https://example.com/file.pdf",
-        %{caption: "Important document"}
-      )
+          "123456789",
+          "https://example.com/file.pdf",
+          %{caption: "Important document"}
+        )
     end
 
     test "send_video/3" do
@@ -1161,10 +795,10 @@ defmodule Lux.Lenses.TelegramBotLensTest do
           "width" => 1280
         }
       }} = TelegramBotLens.send_video(
-        "123456789",
-        "https://example.com/video.mp4",
+          "123456789",
+          "https://example.com/video.mp4",
         %{caption: "Test video"}
-      )
+        )
     end
 
     test "send_audio/3" do
@@ -1202,10 +836,10 @@ defmodule Lux.Lenses.TelegramBotLensTest do
         "from" => %{"id" => 987654321, "is_bot" => true},
         "message_id" => 1
       }} = TelegramBotLens.send_audio(
-        "123456789",
-        "https://example.com/audio.mp3",
+          "123456789",
+          "https://example.com/audio.mp3",
         %{caption: "Test audio"}
-      )
+        )
     end
 
     test "send_location/4" do
@@ -1286,9 +920,9 @@ defmodule Lux.Lenses.TelegramBotLensTest do
       end)
 
       assert {:ok, true} = TelegramBotLens.set_webhook(
-        "https://example.com/webhook",
-        %{max_connections: 40}
-      )
+          "https://example.com/webhook",
+          %{max_connections: 40}
+        )
     end
 
     test "delete_webhook/1" do
